Some testing here are the results:
The main thing blocking an implementation is the point hash function. If
point hash is implemented, the dalek crypto libraries are a beautiful
tool to work with.
The ge_fromfe_frombytes_vartime (ge is group element, fe is field
element) function is what is used to create a point from an array of bytes.
It is disturbingly complicated:

https://github.com/monero-project/research-lab/blob/master/whitepaper/ge_fromfe_writeup/ge_fromfe.pdf
and is implemented in monero:
https://github.com/monero-project/monero/blob/master/src/crypto/crypto-ops.c#L2210
trezor-crypto:
https://github.com/trezor/trezor-crypto/blob/master/ed25519-donna/ed25519-donna-impl-base.c#L640
and mininero:
https://github.com/monero-project/mininero/blob/c5fcee9d8ec8c302bca7fda8ce79b68e20d31c34/mininero.py#L238

Sadly the dalek library does not supply us with functions to do complex
arithmetic such as taking squares roots and exponentiating, so
implementing it will be a giant hassle.

For now I'd like to try out ristretto's signing, but that will obviously
not work out in the long run.

The more sane thing to do is something like:
https://github.com/cryptonote-rust/raw-crypto
This just wraps the crypto functions in monero a rust wrapper. This
should be enough to use them at will.